refactor tags surface maskers
