+++ nilearn/maskers/surface_labels_masker.py
@@ -4,20 +4,19 @@ import warnings
 
 import numpy as np
 from joblib import Memory
-from sklearn.base import BaseEstimator, TransformerMixin
 
 from nilearn import signal
 from nilearn._utils import _constrained_layout_kwargs, fill_doc
-from nilearn._utils.cache_mixin import CacheMixin, cache
+from nilearn._utils.cache_mixin import cache
 from nilearn._utils.class_inspect import get_params
 from nilearn._utils.helpers import is_matplotlib_installed
-from nilearn._utils.tags import SKLEARN_LT_1_6
 from nilearn.maskers._utils import (
     check_same_n_vertices,
     compute_mean_surface_image,
     concatenate_surface_images,
     get_min_max_surface_image,
 )
+from nilearn.maskers.base_masker import _BaseSurfaceMasker
 from nilearn.surface import SurfaceImage
 
 
@@ -47,7 +46,7 @@ def _apply_mask(labels_masker, mask_data, labels_data):
 
 
 @fill_doc
-class SurfaceLabelsMasker(TransformerMixin, CacheMixin, BaseEstimator):
+class SurfaceLabelsMasker(_BaseSurfaceMasker):
     """Extract data from a SurfaceImage, averaging over atlas regions.
 
     .. versionadded:: 0.11.0
@@ -170,33 +169,6 @@ class SurfaceLabelsMasker(TransformerMixin, CacheMixin, BaseEstimator):
         """Return data of label image concatenated over hemispheres."""
         all_labels = [x.ravel() for x in self.labels_img.data.parts.values()]
         return np.concatenate(all_labels)
-
-    def _more_tags(self):
-        """Return estimator tags.
-
-        TODO remove when bumping sklearn_version > 1.5
-        """
-        return self.__sklearn_tags__()
-
-    def __sklearn_tags__(self):
-        """Return estimator tags.
-
-        See the sklearn documentation for more details on tags
-        https://scikit-learn.org/1.6/developers/develop.html#estimator-tags
-        """
-        # TODO
-        # get rid of if block
-        # bumping sklearn_version > 1.5
-        if SKLEARN_LT_1_6:
-            from nilearn._utils.tags import tags
-
-            return tags(surf_img=True, niimg_like=False)
-
-        from nilearn._utils.tags import InputTags
-
-        tags = super().__sklearn_tags__()
-        tags.input_tags = InputTags(surf_img=True, niimg_like=False)
-        return tags
 
     def fit(self, img=None, y=None):
         """Prepare signal extraction from regions.
